Un-split UltraFeed.tsx
